Fetch the hackathon project list instead of always showing the fallback
The "Previous hackathon projects" section has been showing its empty-state
card since it shipped. The cause turned out to be simpler than a rate limit:
fetchHackathonGithubProjects was written and exported, but nothing ever
called it. The hackathon route was a client component taking githubProjects
as a prop that defaulted to an empty array, and an App Router page only ever
receives params and searchParams, so that prop could never be filled. The
empty branch was the only branch it could reach.

The route is now a server component that does the fetch and hands the result
down. The UI moved, unchanged, into components/Hackathon/Hackathon.tsx,
which already held an older copy of the same page that nothing imported.

Also sends GITHUB_TOKEN when it is set, the same optional pattern used in
the data-updated route, since unauthenticated calls only get 60 per hour per
IP. It still works without one. The existing error card is untouched and
still catches a genuinely failing request.
